Update to write and read seq to require input FCB
The read and write sequential SFOS operations now take an FCB and they
update the LBA based on the FCB contents.  They also increment the new
CURRENT RECORD field in the FCB in readiness for the next operation.

To make this work with existing SFCP stuff that does directory
management, I needed to create an internal SFOS directory read function
that does not touch the LBA at all.  This allows directory scanning
functions to update the LBA directly and call the read block code.